[Modify]: 교환글 목록 ui 수정 및 함수 분리 - 교환글 Button 컴퍼넌트 대신 button 태그로 변경 - timeSince 함수 util 폴더로 별도 관리
