+++ src/utils/timeSince.js
@@ -1,0 +1,27 @@
+// 날짜 차이를 계산하는 함수
+export function timeSince(dateToObject) {
+  const date = new Date(Date.parse(dateToObject));
+  const seconds = Math.floor((new Date() - date) / 1000);
+
+  let interval = seconds / 31536000;
+  if (interval > 1) {
+    return Math.floor(interval) + '년 전';
+  }
+  interval = seconds / 2592000;
+  if (interval > 1) {
+    return Math.floor(interval) + '달 전';
+  }
+  interval = seconds / 86400;
+  if (interval > 1) {
+    return Math.floor(interval) + '일 전';
+  }
+  interval = seconds / 3600;
+  if (interval > 1) {
+    return Math.floor(interval) + '시간 전';
+  }
+  interval = seconds / 60;
+  if (interval > 1) {
+    return Math.floor(interval) + '분 전';
+  }
+  return '방금 전';
+}
